feat: check password utility
